refactor: create separate function for recreation of products on the page in the cart
